[ACC-2210] Test against multiple S3 implementations
